Rename missed places to valkey
