+++ glide-core/src/cluster_scan_container.rs
@@ -1,5 +1,5 @@
 /**
- * Copyright GLIDE-for-Redis Project Contributors - SPDX Identifier: Apache-2.0
+ * Copyright Valkey GLIDE Project Contributors - SPDX Identifier: Apache-2.0
  */
 use logger_core::log_debug;
 use nanoid::nanoid;
