+++ python/DEVELOPER.md
@@ -124,7 +124,7 @@ Before starting this step, make sure you've installed all software requirments.
         ```bash
         pytest --asyncio-mode=auto
         ```
-        > **Note:** To run redis modules tests, add -k "test_redis_modules.py".
+        > **Note:** To run redis modules tests, add -k "test_server_modules.py".
 
 -   Install Python development requirements with:
