+++ python/pytest.ini
@@ -2,4 +2,4 @@
 markers =
     smoke_test: mark a test as a build verification testing.
 # TODO: Remove pubsub exclusion after the flakey tests are fixed
-addopts = -k "not redis_modules and not pubsub"
+addopts = -k "not server_modules and not pubsub"
